requirements.txt
Django>=5,<6
Pillow>=10
django-ckeditor>=6.7